% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtd_get_categories.r
\name{dtd_get_categories}
\alias{dtd_get_categories}
\title{Obtain the list of categories}
\usage{
dtd_get_categories(user_token, app_id)
}
\arguments{
\item{user_token}{User token}

\item{app_id}{App ID}
}
\description{
Obtain the list of categories <https://docs.devtodev.com/integration/server-api/labels-api#obtaining-the-list-of-categories>
}
